chore: use eXist-html-templating-lib
This commit removes the ssrq-specific implementation of the templating library. The creater of this commit thinks, that the library was originally copied to the source code, becauce a minir change was made to the functions in templates.xqm (one function was not private any longer). The solution in this commit changes this by using a glue method, which should already be considered as deprecated